+++ DMS/Rpc/monitor/monitor-store-checker/MonitorStoreChecker_StoreCheckingDTO.cs
@@ -14,11 +14,11 @@ namespace DMS.Rpc.monitor.monitor_store_checker
         public long ExternalCounter { get { return External == null ? 0 : External.Count; } }
         public long ImageCounter { get { return Image == null ? 0 : Image.Count; } }
         public long SalesOrderCounter { get { return SalesOrder == null ? 0 : SalesOrder.Count; } }
-        public long RevenueCounter { get { return Revenue == null ? 0 : Revenue.Select(r => r.Value).Sum(); } }
+        public decimal RevenueCounter { get { return Revenue == null ? 0 : Revenue.Select(r => r.Value).Sum(); } }
         internal HashSet<long> Image { get; set; }
         internal HashSet<long> SalesOrder { get; set; }
         internal HashSet<long> Internal { get; set; }
         internal HashSet<long> External { get; set; }
-        internal Dictionary<long, long> Revenue { get; set; }
+        internal Dictionary<long, decimal> Revenue { get; set; }
     }
 }
